// api/src/database/database.types.ts
export type Decimal = ColumnType<string, number | string>;

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export interface Tbacc {
  acceso: number;
  activo: number;
  envia: number;
  fecmod: Date;
  fecnvo: Date;
  id: Generated<number>;
  nomacceso: string;
  usumod: number;
  usunvo: number;
}

export interface Tbaccclases {
  accclase: number;
  acceso: number;
  activo: number;
  clase: number;
  envia: number;
  fecmod: Date;
  fecnvo: Date;
  id: Generated<number>;
  usumod: number;
  usunvo: number;
}

export interface Tbaccesos {
  acceso: number;
  agregar: number;
  eliminar: number;
  envia: number;
  id: Generated<number>;
  imprimir: number;
  modificar: number;
  opcion: number;
  subopcion: number;
  usuario: number;
}

export interface Tbaccperfiles {
  acceso: number;
  agregar: number;
  eliminar: number;
  envia: number;
  id: Generated<number>;
  imprimir: number;
  modificar: number;
  opcion: number;
  perfil: number;
  subopcion: number;
}

export interface Tbagenda {
  correo: string;
  empleado: number;
  envia: Decimal;
  fecha: Date;
  fecmod: Date;
  fecnvo: Date;
  hora: number;
  horarioclase: Generated<number>;
  idagenda: Generated<number>;
  obs: string;
  precita: number;
  producto: number;
  sala: number;
  sesion: number;
  socio: number;
  tomada: Decimal;
  usumod: number;
  usunvo: number;
}

export interface Tbapertura {
  abrir: number;
  envia: number;
  fecmod: Date;
  fecnvo: Date;
  id: Generated<number>;
  usumod: number;
  usunvo: number;
}

export interface Tbaperturas {
  envia: number;
  fecha: Date;
  fecmod: Date;
  fecnvo: Date;
  id: Generated<number>;
  motivo: string | null;
  usumod: number;
  usunvo: number;
}

export interface Tbasistencia {
  autorizo: number;
  clase: number;
  envia: number;
  fecha: Date;
  fecmod: Date;
  fecnvo: Date;
  horario: number;
  id: Generated<number>;
  instructor: number;
  motivo: string;
  retardo: number;
  socio: number;
  usumod: number;
  usunvo: number;
}

export interface Tbbancos {
  banco: Generated<number>;
  cuenta: string;
  envia: number;
  fecmod: Date;
  fecnvo: Date;
  nombanco: string;
  nombre: string;
  rfc: string;
  usumod: number;
  usunvo: number;
}

export interface Tbbdkpc {
  cerrar: number;
  enuso: number;
  envia: number;
  fecmod: Date;
  fecnvo: Date;
  hora: Date;
  minimizar: number;
  parapc: string;
  paraserver: string;
  tiempo: Decimal;
  usumod: number;
  usunvo: number;
}

export interface Tbbloqueados {
  activo: number;
  correobloq: string;
  envia: string;
  fecmod: Date;
  fecnvo: Date;
  id: number;
  usumod: number;
  usunvo: number;
}

export interface Tbcategorias {
  categoria: number;
  color: string;
  enpantalla: number;
  envia: number;
  fecmod: Date;
  fecnvo: Date;
  fotocat: string;
  id: Generated<number>;
  nomcategoria: string;
  principal: number;
  usumod: number;
  usunvo: number;
}

export interface TbcClaveprodser {
  c_claveprodserv: string;
  descripcion: string;
  envia: number;
  fecfinvig: Date;
  fecinivig: Date;
  fecmod: Date;
  fecnvo: Date;
  inclcompl: string;
  inclIEPStr: string;
  inclIVAtr: string;
  usumod: number;
  usunvo: number;
}

export interface TbcClaveunidad {
  c_claveunidad: string;
  descripcion: string;
  envia: number;
  fechafinvig: Date;
  fechainivig: Date;
  fecmod: Date;
  fecnvo: Date;
  nombre: string;
  simbolo: string;
  usumod: number;
  usunvo: number;
}

export interface TbcCodigopostal {
  c_codigopostal: string;
  c_estado: string;
  c_localidad: string;
  c_municipio: string;
  envia: number;
  fecmod: Date;
  fecnvo: Date;
  usumod: number;
  usunvo: number;
}

export interface Tbcfdis {
  cadenaor: string;
  certem: string;
  certsat: string;
  cfdirelacionado: string;
  envia: number;
  factura: number;
  fecmod: Date;
  fecnvo: Date;
  fectim: string;
  sello: string;
  selloem: string;
  serie: string;
  tipodecomprobante: string;
  tipoderelacion: number;
  usumod: number;
  usunvo: number;
  uuid: string;
}

export interface TbcFormapago {
  bancarizado: string;
  bco_emisor_cta_ord_extranjero: string;
  c_formapago: string;
  cta_benenficiario: string;
  cta_ordenante: string;
  descripcion: string;
  envia: number;
  fecmod: Date;
  fecnvo: Date;
  num_oper: string;
  pat_cta_beneficiario: string;
  patron_cta_ordenante: string;
  RFC_emisor_beneficiario: string;
  RFC_emisor_ordenante: string;
  tipo_cadena_pago: string;
  usumod: number;
  usunvo: number;
}

export interface Tbclases {
  activa: number;
  clase: number;
  cntlimite: number;
  cobinsc: number;
  controlhr: number;
  descanual: Decimal;
  descmes: Decimal;
  descqna: Decimal;
  descsem: Decimal;
  descstre: Decimal;
  desctrim: Decimal;
  envia: number;
  fecmod: Date;
  fecnvo: Date;
  id: Generated<number>;
  impticketasist: number;
  limitectes: number;
  nomclase: string;
  pranual: Decimal;
  prinsc: Decimal;
  prmes: Decimal;
  prqna: Decimal;
  prsem: Decimal;
  prstre: Decimal;
  prtrim: Decimal;
  usumod: number;
  usunvo: number;
}

export interface TbcMetodopago {
  c_metodopago: string;
  descripcion: string;
  envia: number;
  fecmod: Date;
  fecnvo: Date;
  usumod: number;
  usunvo: number;
}

export interface TbcMoneda {
  c_moneda: string;
  decimales: string;
  descripcion: string;
  envia: number;
  fecmod: Date;
  fecnvo: Date;
  porcvar: string;
  usumod: number;
  usunvo: number;
}

export interface Tbcobranza {
  cancel: number;
  cliente: number;
  envia: number;
  fecmod: Date;
  fecnvo: Date;
  fecpag: Date;
  formapago: number;
  idfac: number;
  idpag: number;
  importe: Decimal;
  metodopago: string;
  moneda: Decimal;
  noparcialidad: number;
  pago: number;
  referencia: string;
  saldoant: Decimal;
  saldonvo: Decimal;
  tcambiodr: Decimal;
  tipocambio: Decimal;
  usumod: number;
  usunvo: number;
}

export interface Tbconceptos {
  concepto: number;
  envia: number;
  fecmod: Date;
  fecnvo: Date;
  id: Generated<number>;
  nomconcepto: string;
  usumod: number;
  usunvo: number;
}

export interface Tbconsumodieta {
  activo: number;
  consumo: number;
  envia: number;
  fecmod: Date;
  fecnvo: Date;
  id: Generated<number>;
  nomconsumo: string;
  usumod: number;
  usunvo: number;
}

export interface Tbcortes {
  bauchers: Decimal;
  cheques: Decimal;
  compras: Decimal;
  corte: number;
  efecencaja: Decimal;
  efectivo: Decimal;
  entradafin: number;
  entradaini: number;
  envia: number;
  faltante: Decimal;
  fdocaja: Decimal;
  fecha: Date;
  fecmod: Date;
  fecnvo: Date;
  fichasdep: Decimal;
  gastos: Decimal;
  id: Generated<number>;
  ingefec: Decimal;
  obs: string | null;
  salefec: Decimal;
  salidafin: number;
  salidaini: number;
  sobrante: Decimal;
  totingresos: Decimal;
  usuario: number;
  usumod: number;
  usunvo: number;
  vales: Decimal;
  vtascredito: Decimal;
}

export interface TbcRegimenfiscal {
  c_regimenfiscal: string;
  descripcion: string;
  envia: number;
  fecmod: Date;
  fecnvo: Date;
  pfisica: string;
  pmoral: string;
  usumod: number;
  usunvo: number;
}

export interface TbcTasaocuota {
  c_tasaocuotavmax: Decimal;
  c_tasaocuotavmin: string;
  envia: number;
  factor: string;
  fecmod: Date;
  fecnvo: Date;
  impuesto: string;
  rangoofijo: string;
  retencion: string;
  traslado: string;
  usumod: number;
  usunvo: number;
}

export interface TbcTipocadenapago {
  c_tipocadena: string;
  descripcion: string;
  envia: number;
  fecmod: Date;
  fecnvo: Date;
  usumod: number;
  usunvo: number;
}

export interface TbcTipocomprobante {
  c_tipocomprobante: string;
  descripcion: string;
  valormax: string;
}

export interface TbcTiporelacion {
  c_tiporelacion: string | null;
  descripcion: string | null;
  envia: number | null;
  fecfinvig: Date | null;
  fecinivig: Date | null;
  fecmod: Date | null;
  fecnvo: Date | null;
  usumod: number | null;
  usunvo: number | null;
}

export interface TbcUsocfdi {
  appfisica: string;
  appmoral: string;
  c_usoCFDI: string;
  descripcion: string;
  envia: number;
  fecmod: Date;
  fecnvo: Date;
  usumod: number;
  usunvo: number;
}

export interface Tbdescuentos {
  autdes: number;
  cancelado: number;
  corte: number;
  descuento: Decimal;
  envia: number;
  fecha: Date;
  fecmod: Date;
  fecnvo: Date;
  id: Generated<number>;
  iddesc: number;
  idmens: number;
  importe: Decimal;
  motivo: string;
  socio: number;
  ticket: number;
  total: Decimal;
  usuario: number;
  usumod: number;
  usunvo: number;
}

export interface Tbdetcortes {
  corte: number;
  descrip: string;
  envia: number;
  fecmod: Date;
  fecnvo: Date;
  id: Generated<number>;
  importe: Decimal;
  nomsocio: string;
  subtit: number;
  tit: number;
  titulo: string;
  total: number;
  usumod: number;
  usunvo: number;
}

export interface Tbdetdietas {
  activo: number;
  alimentos: string;
  consumo: number;
  dia: number;
  dieta: number;
  envia: number;
  fecmod: Date;
  fecnvo: Date;
  id: Generated<number>;
  usumod: number;
  usunvo: number;
}

export interface Tbdetdietasocio {
  activo: number;
  alimentos: string;
  consumo: number;
  dia: number;
  dieta: number;
  envia: number;
  fecmod: Date;
  fecnvo: Date;
  id: Generated<number>;
  socio: number;
  usumod: number;
  usunvo: number;
}

export interface Tbdetentradasalm {
  cnt: Decimal;
  costo: Decimal;
  envia: number;
  exant: Decimal;
  exnva: Decimal;
  fecmod: Date;
  fecnvo: Date;
  id: Generated<number>;
  ident: number;
  idreg: number;
  producto: number;
  total: Decimal;
  usumod: number;
  usunvo: number;
  venta: Decimal;
}
// ...
export interface DB {
  tbacc: Tbacc;
  tbaccclases: Tbaccclases;
  tbaccesos: Tbaccesos;
  tbaccperfiles: Tbaccperfiles;
  tbagenda: Tbagenda;
  tbapertura: Tbapertura;
  tbaperturas: Tbaperturas;
  tbasistencia: Tbasistencia;
  tbbancos: Tbbancos;
  tbbdkpc: Tbbdkpc;
  tbbloqueados: Tbbloqueados;
  tbc_claveprodser: TbcClaveprodser;
  tbc_claveunidad: TbcClaveunidad;
  tbc_codigopostal: TbcCodigopostal;
  tbc_formapago: TbcFormapago;
  tbc_metodopago: TbcMetodopago;
  tbc_moneda: TbcMoneda;
  tbc_regimenfiscal: TbcRegimenfiscal;
  tbc_tasaocuota: TbcTasaocuota;
  tbc_tipocadenapago: TbcTipocadenapago;
  tbc_tipocomprobante: TbcTipocomprobante;
  tbc_tiporelacion: TbcTiporelacion;
  tbc_usocfdi: TbcUsocfdi;
  tbcategorias: Tbcategorias;
  tbcfdis: Tbcfdis;
  tbclases: Tbclases;
  tbcobranza: Tbcobranza;
  tbconceptos: Tbconceptos;
  tbconsumodieta: Tbconsumodieta;
  tbcortes: Tbcortes;
  tbdescuentos: Tbdescuentos;
  tbdetcortes: Tbdetcortes;
  tbdetdietas: Tbdetdietas;
  tbdetdietasocio: Tbdetdietasocio;
  tbdetentradasalm: Tbdetentradasalm;
  tbdetfacs: Tbdetfacs;
  tbdetoc: Tbdetoc;
  tbdetrutinasocio: Tbdetrutinasocio;
  tbdetsalidasalm: Tbdetsalidasalm;
  tbdettickets: Tbdettickets;
  tbdietas: Tbdietas;
  tbentradasalm: Tbentradasalm;
  tbestatus: Tbestatus;
  tbeventos: Tbeventos;
  tbexistcortes: Tbexistcortes;
  tbfacturas: Tbfacturas;
  tbformaspago: Tbformaspago;
  tbfotosrutinas: Tbfotosrutinas;
  tbgastos: Tbgastos;
  tbgruposprod: Tbgruposprod;
  tbhorarios: Tbhorarios;
  tbhorariosdet: Tbhorariosdet;
  tbhorariossocio: Tbhorariossocio;
  tbhuellas: Tbhuellas;
  tbingresos: Tbingresos;
  tbinstructores: Tbinstructores;
  tblogsocio: Tblogsocio;
  tbmedidas: Tbmedidas;
  tbmensualidades: Tbmensualidades;
  tbmodospago: Tbmodospago;
  tbmotivossal: Tbmotivossal;
  tbnivelrutina: Tbnivelrutina;
  tbnotificaciones: Tbnotificaciones;
  tbocompra: Tbocompra;
  tbpagos: Tbpagos;
  tbparametros: Tbparametros;
  tbparamsat: Tbparamsat;
  tbparterutina: Tbparterutina;
  tbperfiles: Tbperfiles;
  tbprodcancel: Tbprodcancel;
  tbproductos: Tbproductos;
  tbprogramas: Tbprogramas;
  tbproveedores: Tbproveedores;
  tbpublicidad: Tbpublicidad;
  tbreportes: Tbreportes;
  tbrescortes: Tbrescortes;
  tbresenia: Tbresenia;
  tbrutinas: Tbrutinas;
  tbsalidasalm: Tbsalidasalm;
  tbsalidasefec: Tbsalidasefec;
  tbsocios: Tbsocios;
  tbsociosfotos: Tbsociosfotos;
  tbtickets: Tbtickets;
  tbtipodieta: Tbtipodieta;
  tbtiporutina: Tbtiporutina;
  tbunidadesmed: Tbunidadesmed;
  tbusuarios: Tbusuarios;
}